% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_functions.R
\name{add_timepoint_data}
\alias{add_timepoint_data}
\title{Add Timepoint Data}
\usage{
add_timepoint_data(data_path, t1 = "", t2 = "")
}
\arguments{
\item{data_path}{The full path to the Excel file exported from DeepSqueak}

\item{t1}{The start time in the recording (in seconds) to start extracting calls}

\item{t2}{The end time in the recording (in seconds) to stop extracting calls}
}
\value{
An object containing the call data (under time constraints if specified)
}
\description{
Loads in a specific Excel File, and (optional) selects a time subset of data.
}
\examples{
\dontrun{add_timepoint_data(
  data_path = "~/inst/extdata/Example_Mouse_Data.xlsx",
  t1 = 3, t2 = 12
)}

}
