% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_functions.R
\name{add_to_experiment}
\alias{add_to_experiment}
\title{Add Scored Data to Experiment Object}
\usage{
add_to_experiment(experiment, added_data)
}
\arguments{
\item{experiment}{The experiment object}

\item{added_data}{The scored data object to be added to the experiment}
}
\value{
A list object containing statistics and metadata for the entire experiment.
The `groups` and `experimenters` field will auto-populate based on added data using
the `update_experiment()` function.
}
\description{
Adds summarized timepoint data (acquired by running `add_timepoint_data()`,
followed by `score_timepoint_data()`) to the created experiment object.
}
\examples{
\dontrun{add_to_experiment(experiment = experiment, added_data = my_data)}

}
