% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{autosqueakRpipeline}
\alias{autosqueakRpipeline}
\title{Fully-Automatic Experiment Creation}
\usage{
autosqueakRpipeline()
}
\value{
An object containing the full experiment with all data added.
}
\description{
A pipeline function which creates a new experiment object,
and adds data and metadata referenced from an external Google Sheets document.
}
\examples{
expt <- autosqueakRpipeline()

}
