% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_functions.R
\name{create_experiment}
\alias{create_experiment}
\title{Create New Experiment}
\usage{
create_experiment(experiment_name)
}
\arguments{
\item{experiment_name}{The name of the experiment}
}
\value{
A list object containing statistics and metadata for the entire experiment.
The `groups` and `experimenters` field will auto-populate based on added data using
the `update_experiment()` function.
}
\description{
Creates an experiment object which will contain all data and metadata.
This object will be saved and timestamped by date, so if working with this file over
different dates, the object will not be overwritten as a new object will be created
(to preserve backups).
}
\examples{
create_experiment(experiment_name = "My-Project")

}
