% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_functions.R
\name{describe_experiment}
\alias{describe_experiment}
\title{Describe Experiment}
\usage{
describe_experiment(experiment)
}
\arguments{
\item{experiment}{The experiment object to be saved}
}
\value{
A list of information about the experiment
}
\description{
Lists a condensed summary of data stored in the experiment object.
}
\examples{
describe_experiment(experiment = experiment_object)

}
