% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{experiment_pipeline}
\alias{experiment_pipeline}
\title{Full Experiment Creation}
\usage{
experiment_pipeline()
}
\value{
An object containing the full experiment with all data added.
}
\description{
A pipeline function which creates a new experiment object and adds
data by prompting the user
}
\examples{
expt <- experiment_pipeline()

}
