% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing_functions.R
\name{plotCallDataSurface}
\alias{plotCallDataSurface}
\title{3D Surface Plot for Call Datasets}
\usage{
plotCallDataSurface(data_path, res = 60)
}
\arguments{
\item{data_path}{The path to the call data}

\item{res}{The resolution to render surface plot (default is 60)}
}
\value{
3D surface plot for the selected call dataset
}
\description{
Plots interactive 3D plot for density of detected calls
against call length (s) and principal frequency (kHz), using bivariate
kernel density estimation provided via the MASS package
}
\examples{
\dontrun{plotCallDataSurface(data_path = "path")}

}
