% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing_functions.R
\name{plotClusters}
\alias{plotClusters}
\title{3D Call Clusters (Custom Label) Plot}
\usage{
plotClusters(data_path)
}
\arguments{
\item{data_path}{The path to the raw data}
}
\value{
3D Plotly plot
}
\description{
Plots call clusters within a data-point in a 3D Plotly graph, with
principal frequency on the x-axis, mean power on the y-axis, and call length on
the z-axis.
}
\examples{
\dontrun{plotClusters(experiment = experiment, data_id = 1)}

}
