% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing_functions.R
\name{plotContours}
\alias{plotContours}
\title{2D Call Contours}
\usage{
plotContours(data_path)
}
\arguments{
\item{data_path}{The path to the raw data}
}
\value{
2D contour plot
}
\description{
Plots call density against principal frequency and call
length as a contoured heatmap. Histograms for these variables are
displayed across their respective axes.
}
\examples{
\dontrun{plotContours(data_path = "path")}

}
