% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing_functions.R
\name{plotDensityStackedByCustom}
\alias{plotDensityStackedByCustom}
\title{Plot Density Graph Stacked by Custom Labels}
\usage{
plotDensityStackedByCustom(
  data_path,
  graph_title = "Call Distribution Grouped by Custom Category Labels",
  graph_subtitle = "Calls are grouped by custom categories designated in DeepSqueak.",
  chosen_group = c()
)
}
\arguments{
\item{data_path}{The full path to the Excel file exported from DeepSqueak}

\item{graph_title}{The title for the graph (there is a default title if not specified)}

\item{graph_subtitle}{The subtitle for the graph (there is a default subtitle if not specified)}

\item{chosen_group}{Specifying a particular group to be highlighted in the graph}
}
\value{
A ggplot2 visualization of the density graph shown in the viewer window, which can be manually exported.
}
\description{
Plots a density graph, grouped by custom labels set in DeepSqueak, collapsed into a single graph.
}
\examples{
plotDensityStackedByCustom(data_path = "inst/extdata/Example_Mouse_Data.xlsx",
graph_title = "myTitle", graph_subtitle = "myDescription")

}
