% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing_functions.R
\name{plotSunburstAnimals}
\alias{plotSunburstAnimals}
\title{Sunburst Plot for Animal Distribution}
\usage{
plotSunburstAnimals(experiment)
}
\arguments{
\item{experiment}{The experiment object}
}
\value{
Sunburst distribution plotting animal distribution
}
\description{
Plots interactive sunburst plot for animal distribution across
experimental groups. Currently only valid for between-groups studies.
}
\examples{
\dontrun{plotSunburstAnimals(experiment = experiment)}

}
