% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing_functions.R
\name{plotSunburstExperimenters}
\alias{plotSunburstExperimenters}
\title{Sunburst Plot for Experimenter Distribution}
\usage{
plotSunburstExperimenters(experiment)
}
\arguments{
\item{experiment}{The experiment object}
}
\value{
Sunburst distribution plotting experimenter contribution
}
\description{
Plots interactive sunburst plot for experimenter distribution across
experimental groups. Currently only valid for between-groups studies.
}
\examples{
\dontrun{plotSunburstExperimenters(experiment = experiment)}

}
