% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_functions.R
\name{remove_experiment_data}
\alias{remove_experiment_data}
\title{Remove Data from Experiment File}
\usage{
remove_experiment_data(experiment, data_id)
}
\arguments{
\item{experiment}{The experiment object}

\item{data_id}{The number of the data file to be removed}
}
\value{
The experiment file, with the rejected call data removed.
}
\description{
Removes the data object of interest from the experiment according to
a data_id. Removing the data with data_id = 12 removes the 12th dataset added to
the experiment.
}
\examples{
\dontrun{experiment <- remove_experiment_data(experiment = experiment_object, data_id = 12)}

}
