% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_functions.R
\name{save_experiment}
\alias{save_experiment}
\title{Save Experiment Locally}
\usage{
save_experiment(experiment, save_path)
}
\arguments{
\item{experiment}{The experiment object to be saved}

\item{save_path}{The full path where the experiment object will be saved locally}
}
\value{
The experiment object, saved as "[NAME_OF_EXPERIMENT] ([SAVE_DATE]).RData" to the specified location.
}
\description{
Saves the experiment as an .RData file in a specified location.
}
\examples{
\dontrun{save_experiment(experiment = experiment_object, save_path = tempdir())}

}
