% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_functions.R
\name{score_timepoint_data}
\alias{score_timepoint_data}
\title{Score Timepoint Data}
\usage{
score_timepoint_data(data_subset, group, experimenter)
}
\arguments{
\item{data_subset}{The object created in `add_timepoint_data()` which will be scored}

\item{group}{The experimental group (i.e. "Control") these data correspond to}

\item{experimenter}{The experimenter who recorded these results}
}
\value{
A list object containing statistics and metadata for each file.
}
\description{
Transforms data into a list of mean, standard deviation, and range of several call metrics.
}
\examples{
score_timepoint_data(data_subset = my_subsetted_data, group = "Control",
experimenter = "Experimenter 1")

}
