% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_functions.R
\name{semisqueakRpipeline}
\alias{semisqueakRpipeline}
\title{Semi-Automatic Experiment Creation}
\usage{
semisqueakRpipeline()
}
\value{
An object containing the full experiment with all data added.
}
\description{
A pipeline function which creates a new experiment object and adds
data by prompting the user for metadata (for every file that is added).
}
\examples{
expt <- semisqueakRpipeline()

}
