% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SqueakR-package.R
\docType{package}
\name{SqueakR-package}
\alias{SqueakR}
\alias{SqueakR-package}
\title{SqueakR: An Experiment Interface for 'DeepSqueak' Bioacoustics Research}
\description{
Data processing and visualizations for rodent vocalizations exported from 'DeepSqueak'. These functions are compatible with the 'SqueakR' Shiny Dashboard, which can be used to visualize experimental results and analyses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://osimon81.github.io/SqueakR/}
  \item \url{https://github.com/osimon81/SqueakR/}
}

}
\author{
\strong{Maintainer}: Simon Ogundare \email{simon.ogundare@columbia.edu} (\href{https://orcid.org/0000-0002-6243-2779}{ORCID})

}
\keyword{internal}
