% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{squeakrANOVA}
\alias{squeakrANOVA}
\title{Generate ANOVA tables for Detected Calls}
\usage{
squeakrANOVA(experiment, analysis_factor)
}
\arguments{
\item{experiment}{The experiment object}

\item{analysis_factor}{A string representing the factor to analyze between groups}
}
\value{
A bar graph comparing the analysis_factor between groups
}
\description{
Aggregates raw data tables for all data added to experiment object,
and calculates and displays ANOVA statistics.
}
\examples{
\dontrun{squeakrANOVA(experiment = experiment, analysis_factor = "Delta_Freq")}

}
