% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_dashboard.R
\name{squeakRDashboard}
\alias{squeakRDashboard}
\title{Launch SqueakR Dashboard}
\usage{
squeakRDashboard()
}
\value{
A Shiny Dashboard
}
\description{
Run the SqueakR Shiny Dashboard locally to interface with experimental
data.
}
\examples{
\dontrun{squeakRDashboard()}

}
