% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{squeakrSummary}
\alias{squeakrSummary}
\title{Summary Statistics for Experiment}
\usage{
squeakrSummary(experiment)
}
\arguments{
\item{experiment}{The experiment object}
}
\value{
Text which summarizes experiment data split by group
}
\description{
Aggregates raw data for experiment and summarizes descriptive
statistics of the current results, grouped by experimental groups.
}
\examples{
\dontrun{squeakrSTATS(experiment = experiment)}

}
