% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_functions.R
\name{unblind_all_ids}
\alias{unblind_all_ids}
\title{Decode Experiment IDs}
\usage{
unblind_all_ids(experiment)
}
\arguments{
\item{experiment}{The experiment object}
}
\value{
A vector representing the original call file names
}
\description{
Creates a vector of the original call file names, indexed
by the order they are listed in the experiment. This allows experimenters
to unblind themselves to the data they collect
}
\examples{
\dontrun{decode_experiment_ids(experiment)}

}
