% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_functions.R
\name{unblind_data_id}
\alias{unblind_data_id}
\title{Find Matching Experiment ID}
\usage{
unblind_data_id(experiment, filename)
}
\arguments{
\item{experiment}{The experiment object}

\item{filename}{The full name of the file, including the extension}
}
\value{
A number or numbers representing index or indices where that file
appears in the experiment
}
\description{
Finds the index of a dataset matching a supplied file name
in the experiment.
}
\examples{
\dontrun{unblind_data_id(experiment, "my_data1.xlsx")}

}
