% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_functions.R
\name{unblind_data_name}
\alias{unblind_data_name}
\title{Find Matching Experiment Name}
\usage{
unblind_data_name(experiment, id)
}
\arguments{
\item{experiment}{The experiment object}

\item{id}{The dataset number to be unblinded}
}
\value{
The name of the original file, corresponding to the data at
the requested index
}
\description{
Finds the name of a dataset matching a supplied index in the set of data
}
\examples{
\dontrun{unblind_data_name(experiment, 2)}

}
