% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SqueakR-package.R
\docType{package}
\name{SqueakR-package}
\alias{SqueakR}
\alias{SqueakR-package}
\title{SqueakR: An Experiment Superstructure for 'DeepSqueak'}
\description{
A collection of functions intended to produce convenient visualizations for rodent vocalizations exported from 'DeepSqueak'. These visualizations include call frequency-clustered graphs, and density graphs grouped by frequency, custom labels identified in 'DeepSqueak', and call duration.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/osimon81/SqueakR}
}

}
\author{
\strong{Maintainer}: Simon Ogundare \email{simon.ogundare@columbia.edu} (\href{https://orcid.org/0000-0002-6243-2779}{ORCID})

}
\keyword{internal}
