% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_functions.R
\name{update_experiment}
\alias{update_experiment}
\title{Updates Experiment Object Metadata}
\usage{
update_experiment(experiment)
}
\arguments{
\item{experiment}{The experiment object}
}
\value{
A list object containing statistics and metadata for the entire experiment.
The `groups`, `experimenters`, and `animals` fields will auto-populate based on added data using
the `update_experiment()` function.
}
\description{
Auto-populates `groups`, `experimenters`, and `animals` fields in experiment object
by checking experimental data (within the experiment object) for new data.
}
\examples{
\dontrun{update_experiment(experiment)}

}
